All the issues was fixed

+++ src/main/java/com/crossover/techtrial/dto/TopMemberDTO.java
@@ -17,7 +17,7 @@ public class TopMemberDTO {
 	 * @param email
 	 * @param bookCount
 	 */
-	public TopMemberDTO(Long memberId, String name, String email, Integer bookCount) {
+	public TopMemberDTO(Long memberId, String name, String email, Long bookCount) {
 		this.name = name;
 		this.email = email;
 		this.memberId = memberId;
@@ -34,7 +34,7 @@ public class TopMemberDTO {
 
 	private String email;
 
-	private Integer bookCount;
+	private Long bookCount;
 
 	public Long getMemberId() {
 		return memberId;
@@ -60,11 +60,11 @@ public class TopMemberDTO {
 		this.email = email;
 	}
 
-	public Integer getBookCount() {
+	public Long getBookCount() {
 		return bookCount;
 	}
 
-	public void setBookCount(Integer bookCount) {
+	public void setBookCount(Long bookCount) {
 		this.bookCount = bookCount;
 	}
 }
